reworked idling logic of recipe searching
multiblock controllers are notified by the hatches if their inventories changed and the controller may or may not act upon it.
implemented check before trySearchRecipe that should make logic easier to read.
if the previous recipe start failed due to not enough energy, full outputs or if the current inputs did not evaluate into a valid recipe, check for those before re-trying to search a recipe.