enable the recipe logic to make use of the notifiable handlers
recipe logic can wait for changes to the output if it cant process recipes if the output is full. can also wait for changes in the input inventory in case a recipe was searched for but didnt result in a valid recipe.
implemented tests. one simulates a chemical reactor, the other one an electric blast furnace